refactor(servers): migrate database_integration + minecraft_server with thread audit (PR 2d/?, #228)
Two related migrations under #228 PR 2d:

1. database_integration.py -> app/servers/application/database_integration.py
   - Rebuilt on ServersUnitOfWork / ServerRepository (R-15).
   - `update_server_status_sync(server_id, status)` is the
     cross-thread entry registered with MinecraftServerManager.
     It captures the running loop in `initialize()` and bridges
     via asyncio.run_coroutine_threadsafe for true cross-thread
     callers; for same-loop callers (the current code path), it
     schedules the async write fire-and-forget.
   - Sync facade methods (sync_server_states, batch_update,
     get_servers_by_status) are kept as legacy compatibility but
     route through the same loop bridge.

2. minecraft_server.py -> app/servers/application/minecraft_server.py
   - Moved 2,200 LOC file with git-mv (blame preserved on
     subsequent renames).
   - Thread audit: all 22 `_notify_status_change` callsites are
     inside `async def` methods running on the asyncio event loop.
     The registered callback remains sync; bridging happens inside
     DatabaseIntegrationService.update_server_status_sync.
   - Two substantive line changes:
     * Port-conflict block (~line 1287) now routes through
       `ServerRepository.list_by_port(...)` via a late-bound
       `server_repository_factory` (lazy default keeps unit tests
       that instantiate MinecraftServerManager() with no args
       working).
     * `db_session.refresh(server)` at ~line 1148 retained with
       FIXME(#149/#272) marker explaining the transitional
       Session-ownership smell.

Shims at app/services/{database_integration,minecraft_server}.py
re-export the canonical names so existing importers keep working
unchanged.

Lifespan in app/main.py now:
- Late-binds `make_server_repository_from_session` on the
  manager singleton so the new port-conflict path resolves.
- Builds a fresh DatabaseIntegrationService via
  `make_database_integration_service()` and publishes it on
  the module so legacy `database_integration_service` import
  resolves to the lifecycle-aware instance.

Tests:
- test_database_integration_enhanced.py rewritten on top of
  FakeServersUnitOfWork (legacy tests asserted on SQLAlchemy
  internals that no longer exist).
- test_port_validation.py updated to inject a fake
  server_repository_factory.
- Lifecycle integration tests updated similarly so they do not
  fall back into the lazy SqlAlchemy default with a Mock session.
- test_process_persistence.py db_service fixture passes a
  uow_factory.
- test_main_comprehensive.py patches the factory rather than the
  module singleton (lifespan builds fresh).
- Router test patch paths point at the shim
  (`app.services.database_integration.database_integration_service`)
  because the routers still import via the shim.

Acceptance gates:
- `rg db\.query\( app/servers/{domain,application,routers}/` is empty.
- `rg db_session\.query\( app/servers/application/minecraft_server.py`
  is empty.
- `rg asyncio\.get_event_loop\(\)\.run_until_complete app/servers/`
  is empty.
- mypy on app/servers/ + the shims: 0 errors.
- Full `just test` green (1,698 passed, 84 skipped in ~2 min).

Refs #228 (PR 2d/?)
Refs #154
Refs #149 (FIXME for refresh-on-shared-session)
Refs #272 (FIXME for ORM-Server requirement in start_server)
