+++ tests/unit/utils/test_port_validation.py
@@ -1,13 +1,31 @@
-from unittest.mock import Mock, patch
+from unittest.mock import AsyncMock, Mock, patch
 
 import pytest
 
+from app.servers.application.minecraft_server import MinecraftServerManager
 from app.servers.models import Server, ServerStatus, ServerType
-from app.services.minecraft_server import MinecraftServerManager
+
+
+def _make_repo_factory(conflicts):
+    """Build a `server_repository_factory` whose `list_by_port` returns `conflicts`."""
+    repo = Mock()
+    repo.list_by_port = AsyncMock(return_value=conflicts)
+
+    def factory(_db):
+        return repo
+
+    return factory
 
 
 class TestPortValidation:
-    """Test cases for _validate_port_availability method"""
+    """Test cases for `_validate_port_availability` after #228 PR 2d.
+
+    The port-conflict lookup now goes through
+    `ServerRepository.list_by_port(...)` instead of a raw
+    `db_session.query(ServerModel)`. Tests inject a fake
+    `server_repository_factory` on the manager to control the repository
+    response.
+    """
 
     @pytest.fixture
     def manager(self):
@@ -33,39 +51,29 @@ class TestPortValidation:
     async def test_validate_port_availability_no_conflict(
         self, manager, mock_server, mock_db_session
     ):
-        """Test port validation when no conflicts exist"""
-        # Mock database query to return no conflicting servers
-        mock_db_session.query.return_value.filter.return_value.first.return_value = None
-
-        # Mock socket to return connection failed (port available)
+        """Port available when repo returns no conflicts and socket is free."""
+        manager.server_repository_factory = _make_repo_factory(conflicts=[])
         with patch("socket.socket") as mock_socket:
             mock_sock_instance = Mock()
-            mock_sock_instance.connect_ex.return_value = (
-                1  # Connection failed = port available
-            )
+            mock_sock_instance.connect_ex.return_value = 1  # port available
             mock_socket.return_value = mock_sock_instance
 
             available, message = await manager._validate_port_availability(
                 mock_server, mock_db_session
             )
 
-            assert available is True
-            assert "available" in message
+        assert available is True
+        assert "available" in message
 
     @pytest.mark.asyncio
     async def test_validate_port_availability_database_conflict(
         self, manager, mock_server, mock_db_session
     ):
-        """Test port validation when database shows conflicting server"""
-        # Create a mock conflicting server
-        conflicting_server = Mock()
-        conflicting_server.name = "existing-server"
-        conflicting_server.status = ServerStatus.running
-
-        # Mock database query to return conflicting server
-        mock_db_session.query.return_value.filter.return_value.first.return_value = (
-            conflicting_server
-        )
+        """Repo returns a running server using the same port -> conflict."""
+        conflicting = Mock()
+        conflicting.name = "existing-server"
+        conflicting.status = ServerStatus.running
+        manager.server_repository_factory = _make_repo_factory(conflicts=[conflicting])
 
         available, message = await manager._validate_port_availability(
             mock_server, mock_db_session
@@ -79,67 +87,61 @@ class TestPortValidation:
     async def test_validate_port_availability_system_conflict(
         self, manager, mock_server, mock_db_session
     ):
-        """Test port validation when system-level conflict exists"""
-        # Mock database query to return no conflicting servers
-        mock_db_session.query.return_value.filter.return_value.first.return_value = None
-
-        # Mock socket to return connection success (port in use)
+        """No DB conflict but `connect_ex == 0` -> external port-in-use."""
+        manager.server_repository_factory = _make_repo_factory(conflicts=[])
         with patch("socket.socket") as mock_socket:
             mock_sock_instance = Mock()
-            mock_sock_instance.connect_ex.return_value = (
-                0  # Connection success = port in use
-            )
+            mock_sock_instance.connect_ex.return_value = 0
             mock_socket.return_value = mock_sock_instance
 
             available, message = await manager._validate_port_availability(
                 mock_server, mock_db_session
             )
 
-            assert available is False
-            assert "already in use by another process" in message
+        assert available is False
+        assert "already in use by another process" in message
 
     @pytest.mark.asyncio
     async def test_validate_port_availability_no_db_session(self, manager, mock_server):
-        """Test port validation when no database session provided"""
-        # Mock socket to return connection failed (port available)
+        """No db_session -> skip DB check, only socket probe runs."""
         with patch("socket.socket") as mock_socket:
             mock_sock_instance = Mock()
-            mock_sock_instance.connect_ex.return_value = (
-                1  # Connection failed = port available
-            )
+            mock_sock_instance.connect_ex.return_value = 1
             mock_socket.return_value = mock_sock_instance
 
             available, message = await manager._validate_port_availability(
                 mock_server, None
             )
 
-            assert available is True
-            assert "available" in message
+        assert available is True
+        assert "available" in message
 
     @pytest.mark.asyncio
     async def test_validate_port_availability_socket_exception(
         self, manager, mock_server, mock_db_session
     ):
-        """Test port validation when socket operation fails"""
-        # Mock database query to return no conflicting servers
-        mock_db_session.query.return_value.filter.return_value.first.return_value = None
-
-        # Mock socket to raise exception
+        """Socket exception is caught and surfaced as a validation failure."""
+        manager.server_repository_factory = _make_repo_factory(conflicts=[])
         with patch("socket.socket", side_effect=Exception("Socket error")):
             available, message = await manager._validate_port_availability(
                 mock_server, mock_db_session
             )
 
-            assert available is False
-            assert "Port validation failed: Socket error" in message
+        assert available is False
+        assert "Port validation failed: Socket error" in message
 
     @pytest.mark.asyncio
     async def test_validate_port_availability_database_exception(
         self, manager, mock_server, mock_db_session
     ):
-        """Test port validation when database query fails"""
-        # Mock database query to raise exception
-        mock_db_session.query.side_effect = Exception("Database error")
+        """Repository exception is caught and surfaced as a validation failure."""
+        repo = Mock()
+        repo.list_by_port = AsyncMock(side_effect=Exception("Database error"))
+
+        def factory(_db):
+            return repo
+
+        manager.server_repository_factory = factory
 
         available, message = await manager._validate_port_availability(
             mock_server, mock_db_session
@@ -152,16 +154,11 @@ class TestPortValidation:
     async def test_validate_port_availability_starting_server_conflict(
         self, manager, mock_server, mock_db_session
     ):
-        """Test port validation with starting server conflict"""
-        # Create a mock conflicting server that is starting
-        conflicting_server = Mock()
-        conflicting_server.name = "starting-server"
-        conflicting_server.status = ServerStatus.starting
-
-        # Mock database query to return conflicting starting server
-        mock_db_session.query.return_value.filter.return_value.first.return_value = (
-            conflicting_server
-        )
+        """A `starting` server is reported as a conflict (status set passed in)."""
+        conflicting = Mock()
+        conflicting.name = "starting-server"
+        conflicting.status = ServerStatus.starting
+        manager.server_repository_factory = _make_repo_factory(conflicts=[conflicting])
 
         available, message = await manager._validate_port_availability(
             mock_server, mock_db_session
@@ -174,28 +171,28 @@ class TestPortValidation:
     async def test_validate_port_availability_self_exclusion(
         self, manager, mock_server, mock_db_session
     ):
-        """Test that server doesn't conflict with itself"""
-        # Create a mock server with same ID (should be excluded)
-        same_server = Mock()
-        same_server.id = mock_server.id  # Same ID as the server being validated
-        same_server.name = "same-server"
-        same_server.status = ServerStatus.running
-
-        # The database filter should exclude servers with the same ID
-        # We'll mock this by returning None (no conflicts found)
-        mock_db_session.query.return_value.filter.return_value.first.return_value = None
-
-        # Mock socket to return connection failed (port available)
+        """`exclude_id=server.id` keeps the server from conflicting with itself."""
+        # The repository contract is to honour `exclude_id`, so the
+        # fake returns the empty list here — equivalent to "no other
+        # server with this port".
+        manager.server_repository_factory = _make_repo_factory(conflicts=[])
         with patch("socket.socket") as mock_socket:
             mock_sock_instance = Mock()
-            mock_sock_instance.connect_ex.return_value = (
-                1  # Connection failed = port available
-            )
+            mock_sock_instance.connect_ex.return_value = 1
             mock_socket.return_value = mock_sock_instance
 
             available, message = await manager._validate_port_availability(
                 mock_server, mock_db_session
             )
 
-            assert available is True
-            assert "available" in message
+        assert available is True
+        assert "available" in message
+        # And confirm the exclude_id was actually passed through to the
+        # repository call.
+        factory = manager.server_repository_factory
+        repo = factory(mock_db_session)
+        repo.list_by_port.assert_awaited_with(
+            mock_server.port,
+            statuses=[ServerStatus.running, ServerStatus.starting],
+            exclude_id=mock_server.id,
+        )
